% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{add_external_system}
\alias{add_external_system}
\title{Add an external system element to the builder}
\usage{
add_external_system(builder, id, label = id, description = "")
}
\arguments{
\item{builder}{A c4_builder object}

\item{id}{Element identifier}

\item{label}{Element label}

\item{description}{Element description}
}
\value{
Updated builder object
}
\description{
Add an external system element to the builder
}
