% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{add_person}
\alias{add_person}
\title{Add a person element to the builder}
\usage{
add_person(builder, id, label = id, description = "")
}
\arguments{
\item{builder}{A c4_builder object}

\item{id}{Element identifier}

\item{label}{Element label}

\item{description}{Element description}
}
\value{
Updated builder object
}
\description{
Add a person element to the builder
}
