% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{add_system}
\alias{add_system}
\title{Add a system element to the builder}
\usage{
add_system(builder, id, label = id, description = "", technology = "")
}
\arguments{
\item{builder}{A c4_builder object}

\item{id}{Element identifier}

\item{label}{Element label}

\item{description}{Element description}

\item{technology}{Technology stack}
}
\value{
Updated builder object
}
\description{
Add a system element to the builder
}
