% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{build_component}
\alias{build_component}
\title{Build a component diagram from the builder}
\usage{
build_component(builder)
}
\arguments{
\item{builder}{A c4_builder object}
}
\value{
A DiagrammeR graph object
}
\description{
Build a component diagram from the builder
}
