% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{build_container}
\alias{build_container}
\title{Build a container diagram from the builder}
\usage{
build_container(builder)
}
\arguments{
\item{builder}{A c4_builder object}
}
\value{
A DiagrammeR graph object
}
\description{
Build a container diagram from the builder
}
