% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{build_context}
\alias{build_context}
\title{Build a context diagram from the builder}
\usage{
build_context(builder)
}
\arguments{
\item{builder}{A c4_builder object}
}
\value{
A DiagrammeR graph object
}
\description{
Build a context diagram from the builder
}
