% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_c4.R
\name{c4_component_diagram}
\alias{c4_component_diagram}
\title{Create a C4 Component Diagram}
\usage{
c4_component_diagram(
  title = "Component Diagram",
  component = list(),
  external_system = list(),
  relationships = list(),
  theme = "default"
)
}
\arguments{
\item{title}{Character string for diagram title}

\item{component}{List of component elements}

\item{external_system}{List of external system elements}

\item{relationships}{List of relationships between elements}

\item{theme}{Character string for color theme}
}
\value{
A DiagrammeR graph object
}
\description{
Create a C4 Component Diagram
}
