% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_input.R
\name{c4_components_from_tibble}
\alias{c4_components_from_tibble}
\title{Create component elements from a tibble or data frame}
\usage{
c4_components_from_tibble(data)
}
\arguments{
\item{data}{A tibble or data frame with columns: id, label, description, technology}
}
\value{
A list of component elements
}
\description{
Create component elements from a tibble or data frame
}
