% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_input.R
\name{c4_containers_from_tibble}
\alias{c4_containers_from_tibble}
\title{Create container elements from a tibble or data frame}
\usage{
c4_containers_from_tibble(data)
}
\arguments{
\item{data}{A tibble or data frame with columns: id, label, description, technology}
}
\value{
A list of container elements
}
\description{
Create container elements from a tibble or data frame
}
