% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4_demo.R
\name{c4_demo}
\alias{c4_demo}
\title{Demo C4 diagrams}
\usage{
c4_demo()
}
\value{
A list of example C4 diagrams
}
\description{
Creates example C4 diagrams to demonstrate package functionality.
}
\examples{
\dontrun{
demos <- c4_demo()
demos$context # View context diagram
demos$container # View container diagram
demos$component # View component diagram
}
}
