% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_input.R
\name{c4_external_systems_from_tibble}
\alias{c4_external_systems_from_tibble}
\title{Create external system elements from a tibble or data frame}
\usage{
c4_external_systems_from_tibble(data)
}
\arguments{
\item{data}{A tibble or data frame with columns: id, label, description}
}
\value{
A list of external system elements
}
\description{
Create external system elements from a tibble or data frame
}
