% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{c4_find_invalid_rels}
\alias{c4_find_invalid_rels}
\title{Find relationships referencing non-existent elements}
\usage{
c4_find_invalid_rels(elements, relationships)
}
\arguments{
\item{elements}{Named list of elements}

\item{relationships}{List of relationships}
}
\value{
Data frame with invalid relationships
}
\description{
Find relationships referencing non-existent elements
}
