% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_customization.R
\name{c4_load_theme}
\alias{c4_load_theme}
\title{Load a custom theme from file}
\usage{
c4_load_theme(file)
}
\arguments{
\item{file}{File path to load theme from (RDS file)}
}
\value{
A c4_theme object
}
\description{
Load a custom theme from file
}
