% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{c4_save_template}
\alias{c4_save_template}
\title{Save a custom template}
\usage{
c4_save_template(diagram, name, file, parameters = list())
}
\arguments{
\item{diagram}{A diagram object}

\item{name}{Template name}

\item{file}{File path to save template}

\item{parameters}{Named list of parameters used to create the template}
}
\value{
Invisible file path
}
\description{
Save a custom template
}
