% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_customization.R
\name{c4_save_theme}
\alias{c4_save_theme}
\title{Save a custom theme for later use}
\usage{
c4_save_theme(theme, file)
}
\arguments{
\item{theme}{A c4_theme object}

\item{file}{File path to save theme (as RDS)}
}
\value{
Invisible file path
}
\description{
Save a custom theme for later use
}
