% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4r-package.R
\docType{package}
\name{c4r-package}
\alias{c4r}
\alias{c4r-package}
\title{c4r: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
