% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_list}
\alias{check_list}
\title{Check if input is a list}
\usage{
check_list(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{Input to check}

\item{arg}{Argument name for error message}

\item{call}{Calling environment for error reporting}
}
\description{
Check if input is a list
}
\keyword{internal}
