% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_string}
\alias{check_string}
\title{Check if input is a single string}
\usage{
check_string(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{Input to check}

\item{arg}{Argument name for error message}

\item{call}{Calling environment for error reporting}
}
\description{
Check if input is a single string
}
\keyword{internal}
