% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_theme}
\alias{check_theme}
\title{Check if theme is valid}
\usage{
check_theme(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{Input to check}

\item{arg}{Argument name for error message}
}
\description{
Check if theme is valid
}
\keyword{internal}
