% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_c4_edges}
\alias{create_c4_edges}
\title{Create C4 edges}
\usage{
create_c4_edges(relationships)
}
\arguments{
\item{relationships}{List of relationship elements}
}
\value{
Data frame of edges
}
\description{
Create C4 edges
}
\keyword{internal}
