% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_c4_nodes}
\alias{create_c4_nodes}
\title{Create C4 nodes}
\usage{
create_c4_nodes(person, system_or_container, external_system, level)
}
\arguments{
\item{person}{List of person elements}

\item{system_or_container}{List of system/container/component elements}

\item{external_system}{List of external system elements}

\item{level}{Character string indicating diagram level}
}
\value{
Data frame of nodes
}
\description{
Create C4 nodes
}
\keyword{internal}
