% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_node_label}
\alias{create_node_label}
\title{Create formatted node label}
\usage{
create_node_label(label, description, technology)
}
\arguments{
\item{label}{Character string for node label}

\item{description}{Character string for node description}

\item{technology}{Character string for technology}
}
\value{
Character string with formatted label
}
\description{
Create formatted node label
}
\keyword{internal}
