% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generation.R
\name{diagram_from_yaml}
\alias{diagram_from_yaml}
\title{Import diagram elements from YAML}
\usage{
diagram_from_yaml(file)
}
\arguments{
\item{file}{Input YAML file path}
}
\value{
List with elements and relationships
}
\description{
Import diagram elements from YAML
}
