% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_c4_dot}
\alias{export_c4_dot}
\title{Save DOT notation to file}
\usage{
export_c4_dot(diagram, file)
}
\arguments{
\item{diagram}{A DiagrammeR graph object}

\item{file}{Output file path}
}
\value{
Invisible file path
}
\description{
Save DOT notation to file
}
