% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_c4_dot}
\alias{generate_c4_dot}
\title{Generate DOT notation for C4 diagram}
\usage{
generate_c4_dot(title, nodes, edges, theme, level, groups = list())
}
\arguments{
\item{title}{Character string for diagram title}

\item{nodes}{Data frame of nodes}

\item{edges}{Data frame of edges}

\item{theme}{Character string for theme or c4_theme object}

\item{level}{Character string for diagram level}

\item{groups}{List of group objects (optional)}
}
\value{
Character string with DOT notation
}
\description{
Generate DOT notation for C4 diagram
}
\keyword{internal}
