% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generation.R
\name{generate_component_code}
\alias{generate_component_code}
\title{Generate code for a component element}
\usage{
generate_component_code(component)
}
\arguments{
\item{component}{Component element}
}
\description{
Generate code for a component element
}
\keyword{internal}
