% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generation.R
\name{generate_container_code}
\alias{generate_container_code}
\title{Generate code for a container element}
\usage{
generate_container_code(container)
}
\arguments{
\item{container}{Container element}
}
\description{
Generate code for a container element
}
\keyword{internal}
