% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generation.R
\name{generate_diagram_code}
\alias{generate_diagram_code}
\title{Generate diagram creation code}
\usage{
generate_diagram_code(elements, relationships, title, theme, diagram_type)
}
\arguments{
\item{elements}{Named list of elements}

\item{relationships}{List of relationships}

\item{title}{Diagram title}

\item{theme}{Theme name}

\item{diagram_type}{Type of diagram}
}
\description{
Generate diagram creation code
}
\keyword{internal}
