% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generation.R
\name{generate_external_code}
\alias{generate_external_code}
\title{Generate code for an external system}
\usage{
generate_external_code(external)
}
\arguments{
\item{external}{External system element}
}
\description{
Generate code for an external system
}
\keyword{internal}
