% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping.R
\name{generate_group_dot}
\alias{generate_group_dot}
\title{Generate DOT notation for groups (internal helper)}
\usage{
generate_group_dot(groups, theme_colors)
}
\arguments{
\item{groups}{List of group objects}

\item{theme_colors}{Theme color list}
}
\value{
Character string with DOT subgraph notation
}
\description{
Generate DOT notation for groups (internal helper)
}
\keyword{internal}
