% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generation.R
\name{generate_person_code}
\alias{generate_person_code}
\title{Generate code for a person element}
\usage{
generate_person_code(person)
}
\arguments{
\item{person}{Person element}
}
\description{
Generate code for a person element
}
\keyword{internal}
