% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generation.R
\name{generate_rel_code}
\alias{generate_rel_code}
\title{Generate code for a relationship}
\usage{
generate_rel_code(rel)
}
\arguments{
\item{rel}{Relationship element}
}
\description{
Generate code for a relationship
}
\keyword{internal}
