% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generation.R
\name{generate_system_code}
\alias{generate_system_code}
\title{Generate code for a system element}
\usage{
generate_system_code(system)
}
\arguments{
\item{system}{System element}
}
\description{
Generate code for a system element
}
\keyword{internal}
