% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_c4_theme}
\alias{get_c4_theme}
\title{Get C4 theme colors}
\usage{
get_c4_theme(theme)
}
\arguments{
\item{theme}{Character string for theme name}
}
\value{
Named list of colors
}
\description{
Get C4 theme colors
}
\keyword{internal}
