% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_node_style}
\alias{get_node_style}
\title{Get node style based on type}
\usage{
get_node_style(type, colors)
}
\arguments{
\item{type}{Character string for node type}

\item{colors}{Named list of theme colors}
}
\value{
Character string with node style
}
\description{
Get node style based on type
}
\keyword{internal}
