% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{First value}

\item{b}{Default value if a is NULL/empty}
}
\value{
a if not NULL/empty, otherwise b
}
\description{
Null-coalescing operator
}
\keyword{internal}
