% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{print.c4_builder}
\alias{print.c4_builder}
\title{Print method for c4_builder}
\usage{
\method{print}{c4_builder}(x, ...)
}
\arguments{
\item{x}{A c4_builder object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for c4_builder
}
\keyword{internal}
