% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{print.c4_summary}
\alias{print.c4_summary}
\title{Print method for c4_summary}
\usage{
\method{print}{c4_summary}(x, ...)
}
\arguments{
\item{x}{A c4_summary object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for c4_summary
}
