% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generation.R
\name{save_code}
\alias{save_code}
\title{Save generated code to file}
\usage{
save_code(code, file)
}
\arguments{
\item{code}{Character string with R code}

\item{file}{Output file path}
}
\value{
Invisible file path
}
\description{
Save generated code to file
}
