% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{set_theme}
\alias{set_theme}
\title{Set the theme for the diagram}
\usage{
set_theme(builder, theme)
}
\arguments{
\item{builder}{A c4_builder object}

\item{theme}{Theme name or c4_theme object}
}
\value{
Updated builder object
}
\description{
Set the theme for the diagram
}
