% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{set_title}
\alias{set_title}
\title{Set the title for the diagram}
\usage{
set_title(builder, title)
}
\arguments{
\item{builder}{A c4_builder object}

\item{title}{Diagram title}
}
\value{
Updated builder object
}
\description{
Set the title for the diagram
}
