% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{template_data_pipeline}
\alias{template_data_pipeline}
\title{Data pipeline template}
\usage{
template_data_pipeline(
  title = "Data Pipeline",
  sources = c("api", "files"),
  processing = c("etl"),
  storage = c("warehouse")
)
}
\arguments{
\item{title}{Diagram title}

\item{sources}{Character vector of data sources}

\item{processing}{Character vector of processing components}

\item{storage}{Character vector of storage types}
}
\description{
Data pipeline template
}
\keyword{internal}
