% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{template_monolith}
\alias{template_monolith}
\title{Monolith template}
\usage{
template_monolith(
  title = "Monolithic Architecture",
  app_name = "Application",
  database = "postgres",
  cache = NULL
)
}
\arguments{
\item{title}{Diagram title}

\item{app_name}{Application name}

\item{database}{Database type}

\item{cache}{Cache type (optional)}
}
\description{
Monolith template
}
\keyword{internal}
