% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{template_serverless}
\alias{template_serverless}
\title{Serverless template}
\usage{
template_serverless(
  title = "Serverless Architecture",
  functions = c("api", "processor"),
  storage = TRUE,
  database = TRUE
)
}
\arguments{
\item{title}{Diagram title}

\item{functions}{Character vector of function types}

\item{storage}{Include storage}

\item{database}{Include database}
}
\description{
Serverless template
}
\keyword{internal}
