% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{template_three_tier}
\alias{template_three_tier}
\title{Three-tier architecture template}
\usage{
template_three_tier(
  title = "Three-Tier Architecture",
  web_tech = "React",
  app_tech = "Node.js",
  db_tech = "PostgreSQL"
)
}
\arguments{
\item{title}{Diagram title}

\item{web_tech}{Web technology}

\item{app_tech}{Application technology}

\item{db_tech}{Database technology}
}
\description{
Three-tier architecture template
}
\keyword{internal}
