% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_c4r}
\alias{use_c4r}
\title{Scaffold c4r usage in an R package}
\usage{
use_c4r(
  path = ".",
  vignette = TRUE,
  pkgdown = TRUE,
  readme = FALSE,
  template = "three_tier",
  overwrite = FALSE
)
}
\arguments{
\item{path}{Path to the package root. Defaults to the current directory.}

\item{vignette}{Add a starter vignette \code{vignettes/architecture.Rmd}.}

\item{pkgdown}{Add an "Architecture" entry to \verb{_pkgdown.yml}.}

\item{readme}{Add a c4r chunk to \code{README.Rmd}.}

\item{template}{Starter template name (see \code{\link[=c4_list_templates]{c4_list_templates()}}).}

\item{overwrite}{Overwrite existing scaffolding files.}
}
\value{
Invisibly returns the character vector of created/modified files.
}
\description{
Adds an architecture diagram skeleton to the current package: a vignette,
a pkgdown article entry, and (optionally) a chunk in \code{README.Rmd}.
}
\details{
This mirrors the \verb{usethis::use_*} family. It expects to be run from the
root of an R package (where \code{DESCRIPTION} lives), or you can pass \code{path}.
}
\examples{
\dontrun{
# In an R package directory:
use_c4r()
use_c4r(readme = TRUE)
}
}
