% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_c4r_readme}
\alias{use_c4r_readme}
\title{Add a c4r chunk to \code{README.Rmd}}
\usage{
use_c4r_readme(path = ".", overwrite = FALSE)
}
\arguments{
\item{path}{Path to the package root. Defaults to the current directory.}

\item{overwrite}{Overwrite existing scaffolding files.}
}
\value{
Invisibly returns the path to \code{README.Rmd}.
}
\description{
Appends a code chunk to the README that renders an architecture diagram.
If \code{README.Rmd} does not exist, it is created with a minimal header.
}
