% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4r-package.R
\docType{package}
\name{c4r-package}
\alias{c4r}
\alias{c4r-package}
\title{c4r: Create C4 Architecture Diagrams}
\description{
Provides functions to create C4 architecture diagrams (Context, Container, Component, and Code diagrams) using a simple R interface. The package generates diagrams in DOT notation that can be rendered using DiagrammeR or other Graphviz-compatible tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fabiandistler.github.io/c4r/}
  \item \url{https://github.com/fabiandistler/c4r}
  \item Report bugs at \url{https://github.com/fabiandistler/c4r/issues}
}

}
\author{
\strong{Maintainer}: Fabian Distler \email{example@email.com}

}
\keyword{internal}
