% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4r-package.R
\docType{package}
\name{c4r-package}
\alias{c4r}
\alias{c4r-package}
\title{c4r: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fabiandistler/c4r}
  \item Report bugs at \url{https://github.com/fabiandistler/c4r/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
